+++ WebStore/Views/Employees/Index.cshtml
@@ -1,16 +1,14 @@
 ﻿@model IEnumerable<WebStore.Models.Employee>
 
-<!DOCTYPE html>
+@{
+    ViewData["Title"] = "Index";
+}
 
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>Сотрудники</title>
-</head>
-<body>
+<div class="container">
+
     <h3>Сотрудники</h3>
 
-    <table>
+    <table class="table">
         <thead>
             <tr>
                 <th>#</th>
@@ -35,8 +33,7 @@
     </table>
 
     <p>
-        <a href="~/">Домой</a>
+        <a asp-controller="Home" asp-action="Index">Домой</a>
     </p>
 
-</body>
-</html>
+</div>
