+++ WebStore/Views/Shared/_Layout.cshtml
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="">
     <meta name="author" content="">
-    <title>Home | E-Shopper</title>
+    <title>@ViewBag.Title | E-Shopper</title>
 
     @*@await Html.PartialAsync("Partial/_Styles")*@
     <partial name="Partial/_Styles" />
